+++ packages/multichain-account-service/src/index.ts
@@ -18,6 +18,10 @@ export {
   BaseBip44AccountProvider,
   SnapAccountProvider,
   TimeoutError,
+  EVM_ACCOUNT_PROVIDER_NAME,
+  EvmAccountProvider,
+  SOL_ACCOUNT_PROVIDER_NAME,
+  SolAccountProvider,
 } from './providers';
 export { MultichainAccountWallet } from './MultichainAccountWallet';
 export { MultichainAccountGroup } from './MultichainAccountGroup';